API: More closely follow the API used by the reference package.
Basically, when it's not possible to return a randomized shared secret
due to extremely malformed input, RNG failures, or whatever, panic.

This isn't the most friendly thing, but things such as validating input
lengths and not trying to mix and match ParameterSets doesn't feel like
unreasonable demands to place on the calling code.